requirements.txt
streamlit
mediapipe==0.10.9
opencv-python-headless
numpy
pillow
